% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move2ch.r
\name{move2ch}
\alias{move2ch}
\title{Prepare an encounter history in matrix format from telemetry data of tracked individuals.}
\usage{
move2ch(
  study_id = 37350671,
  occasion = "1 month",
  start_cut = Sys.time() - lubridate::years(30),
  end_cut = Sys.time()
)
}
\arguments{
\item{study_id}{numeric. Study-id of a tracking study on Movebank that can be used to download data from \link[=https://movebank.org/]{Movebank}.
Must be publicly accessible.}

\item{occasion}{can be specified in different way compatible with \code{base::seq.Date}. Either as numeric value, which will be taken in days,
or as a character string , containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by an integer and a space, or followed by "s".}

\item{start_cut}{date. Optional cut-off date for the start of the study if some data before that date should be eliminated. Defaults to today-30 years.}

\item{end_cut}{date. Optional cut-off date for the end of the study if some data after that date should be eliminated. Defaults to today.}
}
\value{
Returns a tibble with one row for each individual and one column for each primary occasion (specified by \code{occasion}). The numbers in each matrix cell are the number of GPS locations for that individual in that occasion period.
}
\description{
\code{move2ch} extracts telemetry data from Movebank and converts them into an encounter history for survival estimation where there is
one column for each primary occasion and one row for each individual contained in the data.
}
